requirements.txt
streamlit==1.31.1
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
python-dotenv==1.0.1 